webnn/pool2d_debug.html

<!doctype html>
<meta charset=utf-8>
<title>test pooling operations</title>
<link rel="help" href="https://webmachinelearning.github.io/webnn/#api-mlgraphbuilder-pool2d">
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<script src="./webnn-polyfill.js"></script>
<div id=log></div>
<script type="module">
  'use strict';
  import * as utils from './utils.js';
  let builder;
  setup(() => {
    const context = navigator.ml.createContext();
    builder = new MLGraphBuilder(context);
  })

  test(() => {
    const x = builder.input('x', {type: 'float32', dimensions: [1, 7, 7, 1]});
    const windowDimensions = [4, 4];
    const padding = [1, 1, 1, 1];
    const strides = [2, 2];
    const autoPad = 'explicit';
    const layout = 'nhwc';
    const outputSizes = [4, 4];
    const y = builder.averagePool2d(
        x, {windowDimensions, autoPad, padding, strides, layout, outputSizes});
    const graph = builder.build({y});
    const inputs = {
      'x': new Float32Array([
        1,  2,  3,  4,  5,  6,  7,  8,  9,  10, 11, 12, 13, 14, 15, 16, 17,
        18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34,
        35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49,
      ]),
    };
    const outputs = {y: new Float32Array(utils.sizeOfShape([1, 4, 4, 1]))};
    graph.compute(inputs, outputs);
    const expected = [
      9,
      10.5,
      12.5,
      13.5,
      19.5,
      21,
      23,
      24,
      33.5,
      35,
      37,
      38,
      40.5,
      42,
      44,
      45,
    ];
    assert_array_approx_equals_ulp('averagePool2d', outputs.y, expected, 0);
  }, 'averagePool2d pads outputSizes=[4,4]');
</script>
